Add Pokémon of the Day feature
Added a random Pokémon to appear on the home page. Since this app runs
without a database, a random number seeded with the current date as
string (YYYY-MM-DD) is used to generate the same Pokémon in all
browsers.

Made the border on TypeList the same as in PokemonCard.